css: for h2s within the post body, increased visual attribution
Previously, a h2 element within the post body was equally separated
from the previous paragraph and from the paragraph it was a heading to.

This gave the impression that the element is just a bigger text and
not a heading.  This has been fixed by pushing the preceding paragraph
even higher and bringing the paragraph being headed by the h2 element
closer to the heading itself.